### 2025-06-11 #8392 @editorialbot
[PRE REVIEW]: EchoFlow
**Submitting author:** @erlingdevold (<a href="http://orcid.org/0009-0000-0949-5992">Erling Devold</a>)
**Repository:** https://github.com/erlingdevold/EchoFlow
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
