### 2025-06-12 #8392 @crvernon
:wave: @erlingdevold  - your submission is missing a companion bib file.  Please revisit the submission instructions for JOSS and ensure that your submission meets those requirements before we can move forward with the review.  Thanks.
