+++ paper.md
@@ -15,65 +15,45 @@ affiliations:
     index: 1
 date: 2025-06-10
 doi: 10.5281/zenodo.15634054
+bibliography: bibliography.bib
 ---
 
+# Summary
 
-## Summary
-**EchoFlow** is a three-stage, containerised workflow that turns raw Kongsberg EK80
-sonar files into human-readable images and machine-interpretable attention maps:
+**EchoFlow** is a three-stage, containerised workflow that converts raw Kongsberg EK80 echosounder files into human-readable echograms *and* machine-interpretable attention maps.
 
-1. **Conversion** – decode `.raw` pings, calibrate backscatter with *pyEcholab* [1], and store the result as NetCDF cubes.
-2. **Pre-processing** – contrast-stretch and down-sample echograms as PNG slices.  
-3. **Inference** – apply a self-supervised Vision Transformer (DINO) to generate
-   per-patch attention heatmaps that highlight salient regions such as fish schools
-   or seabed returns.
+1. **Conversion** – raw `.raw` pings are decoded and calibrated to volume-back-scattering strength with *pyEcholab* [@sullivan2018pyecholab].  
+2. **Pre-processing** – echograms are contrast-stretched, down-sampled and tiled as PNGs.  
+3. **Inference** – a Vision Transformer trained with DINO [@caron2021dino] yields per-patch attention heat-maps that highlight fish schools, seabed returns and other salient structures.
 
-The whole pipeline is orchestrated by Docker Compose and verified on commit
-with GitHub Actions.
+Each stage is encapsulated in its own Docker image and orchestrated with Docker Compose; CI ensures that a test file always produces at least one attention map per echogram frequency. The test file can be inspected as artifacts from the CI actions.
 
-## Statement of need
-Marine-acoustics laboratories collect terabytes of sonar each survey but lack a
-single open-source toolchain that
+# Statement of need
 
-* converts multi-frequency data,  
+Marine-acoustics laboratories collect **terabytes** of multi-frequency sonar per survey but lack an open-source tool-chain that
+
+* converts heterogeneous raw formats,  
 * scales from a laptop to an HPC cluster, and  
-* integrates state‑of‑the‑art computer‑vision models.  
+* integrates state-of-the-art computer-vision models.
 
-Existing libraries such as *echopype*  Lee et al. [2] focus on the
-conversion step. While useful for handling raw data, it lacks inference capabilities suitable for deployment on unmanned vehicles.
-EchoFlow fills this gap by chaining conversion → pre-processing → 
-self-supervised inference in one reproducible workflow, enabling rapid analysis, quick modularization and Bring-Your-Own Loop
+Previous work [@lee2024echopype][@sullivan2018pyecholab] addresses the first bullet; EchoFlow fills the remaining gap by chaining **conversion → pre-processing → self-supervised inference** in a single, reproducible workflow. This lowers the barrier for fisheries scientists, marine-robotics engineers and citizen scientists who want modern ML without bespoke pipelines. This pipeline also serves as a foundation for new modern framework incorporated into this science.
 
-## Implementation and architecture
-Each stage lives in its own Docker image and shares data through volume mounts
-(`./data/<stage>`).  A Python watchdog monitors the input directory and triggers
-processing when new files land; continuous‑integration builds the Docker Compose stack and asserts that at least one attention map is produced. 
-Pre-trained DINO weights are downloaded on first use and cached for subsequent runs.
+# Implementation and architecture
[…]
