### 2025-06-12 be1aba3
Create paper.bib
