+++ paper.bib
@@ -1,0 +1,30 @@
+@article{sullivan2018pyecholab,
+  title       = {PyEcholab: An open-source Python-based toolkit to process and visualize echosounder data},
+  author      = {Sullivan, Joshua and Chu, David and Lee, Wu-Jung},
+  journal     = {The Journal of the Acoustical Society of America},
+  volume      = {144},
+  number      = {3},
+  pages       = {1778},
+  year        = {2018},
+  doi         = {10.1121/1.5053981}
+}
+
+@article{lee2024echopype,
+  title       = {Interoperable and scalable echosounder data processing with Echopype},
+  author      = {Lee, Wu-Jung and Setiawan, Landung and Tuguinay, Christian and Mayorga, Emilio and Staneva, Valentina},
+  journal     = {ICES Journal of Marine Science},
+  volume      = {81},
+  number      = {10},
+  pages       = {1941--1951},
+  year        = {2024},
+  doi         = {10.1093/icesjms/fsae133}
+}
+
+@inproceedings{caron2021dino,
+  title       = {Emerging Properties in Self-Supervised Vision Transformers},
+  author      = {Caron, Mathilde and Touvron, Hugo and Misra, Ishan and J{\'e}gou, Herv{\'e} and Mairal, Julien and Bojanowski, Piotr and Joulin, Armand},
+  booktitle   = {Proceedings of the IEEE/CVF International Conference on Computer Vision},
+  pages       = {9639--9650},
+  year        = {2021},
+  doi         = {10.1109/ICCV48922.2021.00951}
+}
