### 2025-06-12 050cdea
add badge
