+++ README.md
@@ -1,6 +1,7 @@
 # Echoflow
 [![CI](https://github.com/erlingdevold/EchoFlow/actions/workflows/ci.yml/badge.svg)](https://github.com/erlingdevold/EchoFlow/actions)
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.15634054.svg)](https://doi.org/10.5281/zenodo.15634054)
+[![status](https://joss.theoj.org/papers/5c9046c3818c08881b51acf6be8d79dc/status.svg)](https://joss.theoj.org/papers/5c9046c3818c08881b51acf6be8d79dc)
 
 >Do you have Terabytes of unprocessed Sonar data? Use this for easy viewing and inference!
 
