### 2025-06-12 6b05971
Update paper.md
Wrong bib key
