### 2025-07-19 #8392 @nkrusch
Hello 👋  @bastienqueste, @malmans2, @patrickcgray, @yueyanglu, and @dspelaez,

would you be interested in reviewing this submission for JOSS?
