### 2025-07-21 #8392 @dspelaez
Hi @nkrusch. Thanks for considering me. Unfortunately, I can't review the submission for JOSS at the moment.
