### 2025-07-30 #8392 @nkrusch
Hi, @devarops, would you be interested in reviewing this submission for JOSS?
