### 2025-07-30 #8392 @nkrusch
Hello, @pnsaevik, would you be interested in reviewing this submission for JOSS?
