### 2025-08-24 #8392 @nkrusch
Hello, @timothyas, would you be interested in reviewing this submission for JOSS?
