### 2025-09-19 #8392 @nkrusch
👋 Hello, @rolker, would you be interested in reviewing [EchoFlow](https://github.com/erlingdevold/EchoFlow) for Journal of Open Source Software (JOSS)?
