### 2025-10-05 #8392 @nkrusch
Hello, @mh-skjelvareid, @emanuelhuber, @castelao, @evanleeturner, and @jessicaaustin:

I am looking for a second reviewer for this submission to JOSS — would any of you be interested?
