### 2025-10-05 #8392 @mh-skjelvareid
@nkrusch I'm sorry, the submission seems interesting, but I just finished a (somewhat lengthy) JOSS review and I don't have time to do another one right now.
