### 2025-10-05 #8392 @emanuelhuber
@nkrusch  I had a quick look at the paper (too short to my point of view, not enough background information on the stuff the code is reading/processing/visualizing) and at the repository (no idea how to install the "tool"/"package"; it is explained how to install docker but for the rest no information). To my point of view, the paper need some more content and the repository need an explanation for dummies on how to install and run it. That would be my review though I do not know exactly what are the expectation of the journal (but if the journal wants to target a larger audience, some improvement are needed).
