### 2025-10-05 #8392 @nkrusch
@emanuelhuber, I appreciate the comments (please take note of them @erlingdevold)!

The Journal of Open Source Software (JOSS) follows a formal peer-review process that is designed to improve the quality of the software submitted. After checking the the submission meets the JOSS [requirements](https://joss.readthedocs.io/en/latest/submitting.html), an editor (me) will look for knowledgeable reviewers (perhaps you?) to conduct the peer review. During the review phase, if a reviewer encounters problems with the submission, reviewer opens issues against the repository for the author to resolve. A more detailed description of the process is given the [review documentation](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) &mdash; in particular, have a look at the [review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html).

The fact that you have naturally followed those steps makes me think you would make an excellent reviewer. Would you like to accept this role?
