### 2025-10-15 62fb2b4
Update paper.bib
doi îs wrong
