+++ paper.bib
@@ -6,7 +6,7 @@
   number      = {3},
   pages       = {1778},
   year        = {2018},
-  doi         = {10.1121/1.5053981}
+  doi         = {10.1121/1.5067860}
 }
 
 @article{lee2024echopype,
