### 2025-11-11 #9240 @nkrusch
@devarops and @subhk, just checking on review progress/status.

We are about a month into the review phase. At this time, it would be good to have attempted a complete inspection of the software. If you encounter problems, open issues against the repository. The review phase proceeds in this manner until all issues are resolved satisfactorily.
