### 2025-11-15 erlingdevold/DINO-pipeline#1 @subhk
Contributing
There seems to be a lack of instructions for contributing to the project
(for example: https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors).

xref: [openjournals/joss-reviews#9240](https://github.com/openjournals/joss-reviews/issues/9240)
