### 2025-11-15 erlingdevold/DINO-pipeline#2 @subhk
Performance
You refer to scaling in the article; could you elaborate on how the package performs and scales on an HPC cluster? Additionally, guidance on setting up and using the package in a cluster environment would be very helpful, as I couldn’t find this information in the README.

xref: [openjournals/joss-reviews#9240](https://github.com/openjournals/joss-reviews/issues/9240)
