### 2025-11-15 erlingdevold/EchoFlow#3 @subhk
Methodology
The package workflow is organized into three stages: Conversion, Pre-processing, and Inference. Please describe how to use the package in each stage, ideally with a short example or recommended usage pattern for each. Also, there is an incomplete sentence at the end of the Setup section of the README file.

xref: [openjournals/joss-reviews#9240](https://github.com/openjournals/joss-reviews/issues/9240)
