### 2025-11-15 erlingdevold/EchoFlow#4 @subhk
Installation instructions
There don’t seem to be any clear instructions on how to install the code. A brief installation guide would be very helpful for newcomers.

xref: [openjournals/joss-reviews#9240](https://github.com/openjournals/joss-reviews/issues/9240)
