### 2025-12-23 #9240 @erlingdevold
> Hi [@erlingdevold](https://github.com/erlingdevold), regarding the 4 open issues:
>
> please give an estimate of when we could expect them to be completed.

Hi, sorry for my inactivity. I will start to close the issues over the holiday. Can expect them completed medio January.
