### 2026-02-24 erlingdevold/DINO-pipeline#5 @nkrusch
Suggested readme enhancements
Ref: [JOSS #9240](https://github.com/openjournals/joss-reviews/issues/9240)

I ran an initial inspection following the readme. I had the following issues with the documentation.

1. [x] Clarify what "Populate input" does

   Please add a plaintext description of what this command does, so it is known in advance.

2. [x] Prevent this git-related Docker build issue.

    Initially I did not notice that the repository had submodules. When I ran Docker build, it errored with pip error.
    The actual issue was that I had not cloned the submodules.[^1] This fix is not  obvious from the observed error.
    I recommend addressing it by including in the readme clearly that repository should be cloned with submodules.[^2]

3. [x] I am able to run the container, it remains unclear what should happen next.

    I opened the browser at the Dashboard `http://0.0.0.0:8050/`.
    Under `data/input` I have 105 MB input file from aws from populate input step.
    However the dash is empty and does not seem to process anything (possible functionality issue).

[^1]: I had run `git submodule sync --recursive` during populate input, but it does not initialize.
[^2]: After clone, running `git submodule update --init --recursive` will fix uninitialized submodules.
