+++ paper.bib
@@ -1,5 +1,5 @@
 @article{sullivan2018pyecholab,
-  title       = {PyEcholab: An open-source Python-based toolkit to process and visualize echosounder data},
+  title       = {PyEcholab: An open-source {P}ython-based toolkit to process and visualize echosounder data},
   author      = {Sullivan, Joshua and Chu, David and Lee, Wu-Jung},
   journal     = {The Journal of the Acoustical Society of America},
   volume      = {144},
@@ -10,7 +10,7 @@
 }
 
 @article{lee2024echopype,
-  title       = {Interoperable and scalable echosounder data processing with Echopype},
+  title       = {Interoperable and scalable echosounder data processing with {E}chopype},
   author      = {Lee, Wu-Jung and Setiawan, Landung and Tuguinay, Christian and Mayorga, Emilio and Staneva, Valentina},
   journal     = {ICES Journal of Marine Science},
   volume      = {81},
