### 2026-02-25 03606f6
Add contribution info to echoflow
