### 2026-02-25 ddee3cf
Update README, paper softens claim
