+++ paper.md
@@ -33,12 +33,12 @@ Each stage is encapsulated in its own Docker image and orchestrated with Docker 
 
 Marine-acoustics research collect **terabytes** of multi-frequency sonar per survey but lack an open-source tool-chain that
 
-* converts heterogeneous raw formats,  
-* scales from a laptop to an HPC cluster, and  
+* converts heterogeneous raw formats,
+* scales from a laptop to HPC machines, and
 * integrates state-of-the-art computer-vision models.
 
-Previous work [@lee2024echopype;@sullivan2018pyecholab] addresses the first bullet; EchoFlow fills the remaining gap by chaining **conversion → pre-processing → self-supervised inference** in a single, reproducible workflow. 
-This lowers the barrier for fisheries scientists, marine-robotics engineers and citizen scientists who want modern ML without bespoke pipelines. This pipeline also serves as a foundation for new modern framework incorporated into this science.
+Previous work [@lee2024echopype][@sullivan2018pyecholab] addresses the first bullet; EchoFlow fills the remaining gap by chaining **conversion → pre-processing → self-supervised inference** in a single, reproducible workflow.
+This lowers the barrier for fisheries scientists, marine-robotics engineers and citizen scientists who want modern ML without bespoke pipelines. The containerised design also facilitates deployment on shared compute environments via tools such as Singularity/Apptainer. This pipeline also serves as a foundation for new modern framework incorporated into this science.
 
 # Implementation and architecture
 
