### 2026-02-25 erlingdevold/DINO-pipeline#1 @erlingdevold
Added `CONTRIBUTING.md`. Covers bug reports, feature suggestions, setup and so on. ref: 03606f65a3393ff35489e3697b56c04d66273df9

xref: [openjournals/joss-reviews#9240](https://github.com/openjournals/joss-reviews/issues/9240)
