### 2026-02-25 38ed8bb
update README for HPC documentation
