### 2026-02-25 erlingdevold/DINO-pipeline#2 @erlingdevold
Clarified the HPC claim in  `paper.md` and the README. The scaling is intra-node parallelism via process pools. There is no built-in cluster or scheduler integration. The README now has a "Performance / parallelism" section making this explicit, and the paper bullet reads "scales from a laptop to HPC machines" to reflect that more cores on a single node (including HPC nodes) gives proportional throughput gains.

xref: [openjournals/joss-reviews#9240](https://github.com/openjournals/joss-reviews/issues/9240)
