### 2026-02-25 erlingdevold/DINO-pipeline#3 @erlingdevold
Added a per-stage usage table to the README showing how to run each stage individually (`docker compose up --build raw`, `preprocessing`,`infer`). Also removed the truncated sentence that ended mid-way.

xref: [openjournals/joss-reviews#9240](https://github.com/openjournals/joss-reviews/issues/9240)
