### 2026-02-25 erlingdevold/DINO-pipeline#4 @erlingdevold
Added an Installation section to the README with prerequisites, explicit `git clone --recurse-submodules` instructions, and a pointer to the per-stage `requirements.txt` files for non-Docker use.

xref: [openjournals/joss-reviews#9240](https://github.com/openjournals/joss-reviews/issues/9240)
