### 2026-02-25 erlingdevold/DINO-pipeline#5 @erlingdevold
Fixed two things:
(1) Added a  to the "Populate input" section explaining that it fetches a public NOAA EK80 test file (~105MB).
(2) Fixed `docker-compose.yml`; the monitor service now has volume mounts for `OUTPUT_DIR`:`./data/preprocessing:/app/output` and  `PP_OUTPUT_DIR`:`./data/inference:/app/pp_output`, so file counts will be non-zero once the pipeline runs.
Also added a "Monitor dashboard" paragraph clarifying that `http://localhost:8050` is a progress monitor , not an inference viewer (currently), and listing where actual outputs land.
