### 2026-02-25 erlingdevold/DINO-pipeline#7 @nkrusch
Feedback on Software Paper
Ref: https://github.com/openjournals/joss-reviews/issues/9240

I have the following comments and suggestions

### Summary

* Briefly introduce _"raw Kongsberg EK80 echosounder files"_:
  - I wonder is this the standard file format in marine acoustics, or is it one of many formats in use?
    (The former implies broader generality, and the latter means there are some limitations by format)
  - I assume these files are probably very large, but only after reading the Statement of the Need. Is the large size of these files part of the challenge that justifies the need for EchoFlow?

* pyEcholab should be PyEcholab (capitalize P)

* _"tiled as PNGs."_ Presumably PNG images? please clarify.

* _"fish schools, seabed returns and other salient structures."_
  Oxford comma: "seabed returns, and…"

* _"…Docker Compose; CI"_
  I would end the sentence at _"Docker Compose."_ Then, _"Continuous Integration (CI)…"_
 You will want to define CI once before use.

### Statement of the need

* How is the problem of processing echosounder files solved currently, without EchoFlow?
  This could also be discussed in the summary section.

* _"Marine-acoustics research collect terabytes of multi-frequency sonar per survey…"_
  - Consider a different noun, like Marine-acoustics _researchers_; the current choice does not work.
   (_who_ collects the terabytes of sonar and needs a tool-chain?)
  - "sonar" -> sonar data?

### Illustrative example

* Actually add the generated illustrations (38000_debug.jpg and 70000.png) as a figure.

* The line `s3://noaa-wcsd-pds/data/raw/Bell_...` is too long to fit on the page.
  Perhaps the there is a way to shorten it, or line wrap (at slash or hyphen) around `EK80/Hake-D20230811-`?

### Missing sections

Ref: https://joss.readthedocs.io/en/pre-2026-submission/paper.html#what-should-my-paper-contain

* Mention (if applicable) a representative set of past or ongoing research projects using the software and recent scholarly publications enabled by it.
* Acknowledgement of any financial support. If there is no financial support, then state that.
