### 2026-02-27 erlingdevold/DINO-pipeline#7 @erlingdevold
@nkrusch  Thank you for great feedback! Skimmed through it and hope to address it from week 12 and onwards. Currently busy with another submission deadline!
