### 2026-03-10 erlingdevold/DINO-pipeline#8 @nkrusch
Python nested dependency on HDF5 and netCDF headers
I am using macOS 15.7.3 (arm) machine.

When installing Python dependencies, I see a non-terminating error message, coming from the dependency:

```
Collecting netCDF4==1.6.5 (from -r raw_consumer/requirements.txt (line 83))

Package hdf5 was not found in the pkg-config search path.
      Perhaps you should add the directory containing `hdf5.pc'
      to the PKG_CONFIG_PATH environment variable
      Package 'hdf5' not found
      reading from setup.cfg...

HDF5_DIR environment variable not set, checking some standard locations ..
⋮
(omitted)
⋮
ValueError: did not find HDF5 headers
```

I can install the required headers (with, e.g., homebrew `brew install hdf5 netcdf`) which resolves the setup issue.

However,  I do recommend putting an advance notice of the HDF5 and netcdf headers requirement in the readme.

ref: https://github.com/openjournals/joss-reviews/issues/9240
