### 2026-03-10 erlingdevold/DINO-pipeline#9 @nkrusch
Software documentation
All the repository documentation is currently in the readme. This is fine for small projects, but eventually you will need more space and a more robust setup.

Here are some suggestions for possible setups for documentation:
- [read the docs](https://docs.readthedocs.com/platform/stable/tutorial/index.html)
- [Jekyll with GitHub Pages](https://docs.github.com/en/pages/setting-up-a-github-pages-site-with-jekyll)
- [docusaurus](https://docusaurus.io/docs/category/getting-started)
- [material mkdocs](https://squidfunk.github.io/mkdocs-material/) (etc.)

I recommend that you setup a minimal workflow and website for documentation. It can just contain a copy of the readme for now; then expand from there gradually. For example, a page of examples of echograms and attention maps would be useful to visualize the software capabilities.

Ref: [openjournals/joss-reviews/issues/9240](https://github.com/openjournals/joss-reviews/issues/9240)
