### 2026-03-14 erlingdevold/DINO-pipeline#10 @subhk
Performance issues
Please **do not close the issue** if you think you have answered my questions, because that makes it difficult to continue the conversation.

My question still has not been answered:
https://github.com/erlingdevold/EchoFlow/issues/2

The package is described as designed to analyze terabytes (TB) of SODAR data. My question is very simple: if I have 10 TB of data and I want to analyze it using 10 processors, how long will it take — seconds, minutes, hours, or days? Also, if I increase the number of processors from 10 to 20, will I see any speed-up?

This is important because it allows users to make a ballpark estimate when designing their experiment.

Ref: https://github.com/openjournals/joss-reviews/issues/9240
