### 2026-03-14 erlingdevold/DINO-pipeline#11 @subhk
Absence of a test suite
I realized that the package contains no unit tests. The sole CI mechanism (.github/workflows/ci.yml) is an end-to-end smoke test that downloads a real sonar file from NOAA's public S3 bucket, runs the full Docker stack, and checks whether PNG files appear within 5 minutes. This is not sufficient for several reasons -

- It creates an empty checkpoint.pth (touch inference/checkpoint.pth), forcing the inference stage through an error-handling fallback that silently switches model architectures

- It verifies file existence but not file correctness

- Individual processing functions (generate_freq_sv_ds, sv_to_jpg, get_attention_maps) have no isolated tests

The authors should provide unit tests for each processing function with synthetic or small reference datasets, and integration tests that verify output correctness against known references.

Ref: https://github.com/openjournals/joss-reviews/issues/9240
