### 2026-03-14 erlingdevold/DINO-pipeline#12 @subhk
Pipeline destroys input data
Files: preprocessing/preprocessing.py:159-167

The mark_as_processed function deletes source .nc files after processing:

def mark_as_processed(file: Path):
    try:
        file.unlink()
        log.info(f"Deleted file {file}")
        marker_file = file.with_suffix(".processed")
        marker_file.touch()
        log.info(f"Created marker file {marker_file}")
    except Exception as e:
        log.error(f"Error deleting file {file} or creating marker: {e}")

Deletion as a state-tracking mechanism is fundamentally unsafe. The authors should adopt a non-destructive approach - either move completed files to an archive directory or use marker files alongside the originals without deleting the originals.

Ref: https://github.com/openjournals/joss-reviews/issues/9240
