### 2026-03-14 erlingdevold/DINO-pipeline#14 @subhk
Diagnostic logs are destroyed on process restart
Files: watchdog.py:23, raw_consumer/raw.py:18, inference/inspect_attention.py:15

Three of the four pipeline stages configure logging with filemode="w", which truncates the log file on initialization. Each time a container restarts, all logs from the previous run are permanently destroyed. (preprocessing/preprocessing.py is unaffected — it uses the default filemode="a".)

The entrypoint.sh shared by all stages runs python main.py followed by python stat.py (the watchdog). Within a single run, the transition from main.py to stat.py does not re-truncate the log, because Python's logging.basicConfig is a no-op once the root logger is already configured. However, any container restart — whether due to a crash, redeployment, or manual restart — truncates the log file and destroys all prior diagnostic output.

In a pipeline with data-loss bugs, the inability to perform post-hoc forensic analysis is a critical gap. The authors should use filemode="a" and implement log rotation.

---
  Key changes:
  - Removed preprocessing/preprocessing.py from the file list and noted it's unaffected
  - Corrected the truncation mechanism: it happens on container restart, not via the main → stat.py import transition
  - Explained why basicConfig is a no-op within a single run

Ref: https://github.com/openjournals/joss-reviews/issues/9240
