### 2026-03-14 erlingdevold/DINO-pipeline#15 @subhk
Inference stage lacks parallelization
I was under the impression that all three stages (Conversion, Pre-processing, and Inference) are parallelized. But going through the code, I realized that the Inference stage is not parallelized.

File: inference/inspect_attention.py:149-156

The raw conversion and preprocessing stages both use ProcessPoolExecutor to process files in parallel (raw_consumer/raw.py:156, preprocessing/preprocessing.py:242). The inference stage, however, processes all files sequentially via nested for loops -

  def consume_dir(input_dir: Path, output_dir: Path):
      files_to_compute = reduce_files_to_diff(input_dir, output_dir)
      for folder in files_to_compute:
          for file in folder.glob("*.png"):
              infer_attention_heads(file, ...)  # sequential, one at a time

The author should explain whether the particular stage is the most computationally expensive or not.

This gives a **false impression** that the code is parallelized, as mentioned in the README.md file:

 **EchoFlow uses process pools within each stage to parallelise work across .raw files.**

Ref: https://github.com/openjournals/joss-reviews/issues/9240
