### 2026-03-14 erlingdevold/DINO-pipeline#16 @subhk
Misleading HPC claim
The paper claims EchoFlow "scales from a laptop to HPC machines". HPC conventionally refers to multi-node, distributed computing - typically involving job schedulers (Slurm, PBS), inter-node communication (MPI), or distributed task frameworks. EchoFlow has none of these. Its parallelism is limited to ProcessPoolExecutor within a single container on a single node, and the inference stage is not parallelized at all.

The authors should replace "HPC machines" with accurate terminology, such as "multi-core machines".

Ref: https://github.com/openjournals/joss-reviews/issues/9240
