### 2026-03-15 erlingdevold/DINO-pipeline#17 @subhk
hardcoded worker count
Files: raw_consumer/raw.py:150, preprocessing/preprocessing.py:236

max_workers=12 is hardcoded. In resource-constrained Docker containers, this can exceed available CPU. The worker count should be configurable via an environment variable, with a sensible default (e.g., os.cpu_count()).

Ref: https://github.com/openjournals/joss-reviews/issues/9240
