### 2026-03-15 erlingdevold/DINO-pipeline#19 @subhk
minor issues
1. inspect_attention.py — model reloaded per image

  File: inference/inspect_attention.py:121

infer_attention_heads calls setup_device_and_model() for every image file, reloading model weights from disk each time. The model and device should be initialized once and passed as arguments.

2. inspect_attention.py — data Variable is dead code

    File: inference/inspect_attention.py:150-161

consume_dir initializes data = None, never reassigns it, then logs "No data was processed" unconditionally. This suggests incomplete implementation.

3. Duplicate import gc

  File: preprocessing/preprocessing.py:194, 206

  The gc module is imported twice.

Ref: https://github.com/openjournals/joss-reviews/issues/9240
