### 2026-03-16 erlingdevold/DINO-pipeline#11 @nkrusch
@erlingdevold, let me comment on this, because I already marked this item satisfactory.

My satisfactory judgement was based on the presence of the CI pipeline; however, I concur with the arguments made by @subhk. The current CI does indeed perform a crude test. I further agree that adding unit tests would improve the development workflow, code quality, and protect against breaking changes in the future.

Beyond meeting the JOSS criteria, I encourage you to add unit tests for the long-term benefit they offer.
