### 2026-03-17 c030b8e
Fix critical bugs: mask NameError, double mark_as_processed, data deletion optional (default false), append log, dup. imports
Adresses #18, #13 #12 #13, #19.3, 17, 19,15.
	modified:   README.md
	modified:   docker-compose.yml
	modified:   inference/inspect_attention.py
	modified:   preprocessing/preprocessing.py
	modified:   raw_consumer/raw.py
	modified:   raw_consumer/requirements.txt
	modified:   watchdog.py
	modified:   paper.md
