+++ paper.md
@@ -22,23 +22,28 @@ bibliography: paper.bib
 # Summary
 
 **EchoFlow** is a three-stage, containerised workflow that converts raw Kongsberg EK80 echosounder files into human-readable echograms *and* machine-interpretable attention maps.
+The EK80 is the dominant wideband echosounder from Kongsberg, the leading manufacturer in fisheries acoustics; surveys with these instruments routinely produce terabytes of multi-frequency data.
 
-1. **Conversion** – raw `.raw` pings are decoded and calibrated to volume-back-scattering strength with *pyEcholab* [@sullivan2018pyecholab].  
-2. **Pre-processing** – echograms are contrast-stretched, down-sampled and tiled as PNGs.  
-3. **Inference** – a Vision Transformer trained with DINO [@caron2021dino] yields per-patch attention heat-maps that highlight fish schools, seabed returns and other salient structures.
+1. **Conversion** – raw `.raw` pings are decoded and calibrated to volume-back-scattering strength with *PyEcholab* [@sullivan2018pyecholab].
+2. **Pre-processing** – echograms are contrast-stretched, down-sampled and tiled as PNG images.
+3. **Inference** – a Vision Transformer trained with DINO [@caron2021dino] yields per-patch attention heat-maps that highlight fish schools, seabed returns, and other salient structures.
 
-Each stage is encapsulated in its own Docker image and orchestrated with Docker Compose; CI ensures that a test file always produces at least one attention map per echogram frequency. The test file can be inspected as artifacts from the CI actions.
+Each stage is encapsulated in its own Docker image and orchestrated with Docker Compose.
+Continuous Integration (CI) ensures that a test file always produces at least one attention map per echogram frequency. The test file can be inspected as artifacts from the CI actions.
 
 # Statement of need
 
-Marine-acoustics research collect **terabytes** of multi-frequency sonar per survey but lack an open-source tool-chain that
+Marine-acoustics researchers collect **terabytes** of multi-frequency sonar data per survey but lack an open-source tool-chain that
 
 * converts heterogeneous raw formats,
 * scales from a laptop to HPC machines, and
 * integrates state-of-the-art computer-vision models.
 
+Today the typical workflow relies on ad-hoc scripts combining PyEcholab or Echopype for format conversion, followed by manual preprocessing and separate ML tooling—or on proprietary software such as Echoview.
+No single open-source pipeline bridges raw data and ML-ready outputs.
+
 Previous work [@lee2024echopype][@sullivan2018pyecholab] addresses the first bullet; EchoFlow fills the remaining gap by chaining **conversion → pre-processing → self-supervised inference** in a single, reproducible workflow.
-This lowers the barrier for fisheries scientists, marine-robotics engineers and citizen scientists who want modern ML without bespoke pipelines. The containerised design also facilitates deployment on shared compute environments via tools such as Singularity/Apptainer. This pipeline also serves as a foundation for new modern framework incorporated into this science.
+This lowers the barrier for fisheries scientists, marine-robotics engineers, and citizen scientists who want modern ML without bespoke pipelines. The containerised design also facilitates deployment on shared compute environments via tools such as Singularity/Apptainer. This pipeline also serves as a foundation for incorporating modern frameworks into marine science.
 
 # Implementation and architecture
 
@@ -50,9 +55,9 @@ Images are multi-platform (`linux/amd64`, `linux/arm64`).
 
 ```bash
[…]
