+++ inference/inspect_attention.py
@@ -12,11 +12,12 @@ log = os.getenv("LOG_DIR", ".")
 logging.basicConfig(
     filename=Path(log) / "raw.log",
     level=logging.DEBUG,
-    filemode="w",
+    filemode="a",
     format="%(asctime)s - %(message)s",
 )
 
-import functools
+import functools  # noqa: E402
+
 
 def log_errors(func):
     @functools.wraps(func)
@@ -30,99 +31,149 @@ def log_errors(func):
     return wrapper
 
 
-import numpy as np
-import torch
-from PIL import Image, ImageOps
-import torchvision.transforms as pth_transforms
-import torch.nn as nn
-import matplotlib.pyplot as plt
+import numpy as np  # noqa: E402
+import torch  # noqa: E402
+from PIL import Image  # noqa: E402
+import torchvision.transforms as pth_transforms  # noqa: E402
+import torch.nn as nn  # noqa: E402
+import matplotlib.pyplot as plt  # noqa: E402
 
 
-import vision_transformer as vits
-def setup_device_and_model(arch='vit_tiny', patch_size=16, pretrained_weights=None, checkpoint_key='teacher'):
-    device = torch.device("cpu")
+import vision_transformer as vits  # noqa: E402
+
+DEVICE = torch.device(
+    os.getenv("DEVICE", "cuda" if torch.cuda.is_available() else "cpu")
+)
+BATCH_SIZE = int(os.getenv("BATCH_SIZE", 4))
+
+
+def setup_device_and_model(
+    arch="vit_tiny", patch_size=16, pretrained_weights=None, checkpoint_key="teacher"
+):
+    device = DEVICE
     model = vits.__dict__[arch](patch_size=patch_size, num_classes=0).eval().to(device)
     for p in model.parameters():
         p.requires_grad = False
 
     if pretrained_weights and os.path.isfile(pretrained_weights):
         try:
-            state_dict = torch.load(pretrained_weights, map_location="cpu",weights_only=False).get(checkpoint_key, {})
-            state_dict = {k.replace("module.", "").replace("backbone.", ""): v for k, v in state_dict.items()}
+            state_dict = torch.load(
+                pretrained_weights, map_location=device, weights_only=False
+            ).get(checkpoint_key, {})
+            state_dict = {
+                k.replace("module.", "").replace("backbone.", ""): v
+                for k, v in state_dict.items()
+            }
             model.load_state_dict(state_dict, strict=False)
-        except Exception as e:
+        except Exception:
             # pull from timm
-            if  patch_size == 16:
+            if patch_size == 16:
                 url = "dino_deitsmall16_pretrain/dino_deitsmall16_pretrain.pth"
             elif patch_size == 8:
                 url = "dino_deitsmall8_300ep_pretrain/dino_deitsmall8_300ep_pretrain.pth"  # model used for visualizations in our paper
             else:
                 raise ValueError(f"Invalid patch size: {patch_size}")
 
-            model = vits.__dict__['vit_small'](patch_size=patch_size, num_classes=0).eval().to(device)
-            state_dict = torch.hub.load_state_dict_from_url(url="https://dl.fbaipublicfiles.com/dino/" + url, map_location="cpu", weights_only=False)
+            model = (
+                vits.__dict__["vit_small"](patch_size=patch_size, num_classes=0)
+                .eval()
+                .to(device)
+            )
+            state_dict = torch.hub.load_state_dict_from_url(
+                url="https://dl.fbaipublicfiles.com/dino/" + url,
+                map_location=device,
+                weights_only=False,
+            )
[…]
