### 2026-03-17 64d7787
add unit tests for issue #11.
	modified:   .github/workflows/ci.yml
	new file:   tests/__init__.py
	new file:   tests/conftest.py
	new file:   tests/test_inference.py
	new file:   tests/test_preprocessing.py
	new file:   tests/test_raw.py
