### 2026-03-17 8ca0665
Add paper fixes, HDF5 docs, benchmark script, and mkdocs site
Address #7,#8,#9,#10,#16
	new file:   .github/workflows/docs.yml
	modified:   README.md
	new file:   docs/gallery.md
	new file:   docs/index.md
	new file:   docs/installation.md
	new file:   docs/usage.md
	new file:   mkdocs.yml
	modified:   paper.md
	new file:   benchmark.py
