### 2026-03-24 erlingdevold/DINO-pipeline#7 @nkrusch
Issue #16 is also related to software paper.
