### 2026-04-03 #9240 @nkrusch
@erlingdevold please provide a status update.

I recommend prioritizing the software paper, then working out the technical issues.
