### 2026-04-04 erlingdevold/DINO-pipeline#12 @erlingdevold
Fixed in c030b8e.
KEEP_INTERMEDIATES env var now preserves .nce files, deletion is opt-in. `preprocessing/preprocessing.py`
