### 2026-04-04 erlingdevold/DINO-pipeline#13 @erlingdevold
Fixed in c030b8e. `mark_as_processed` is called exactly once, in `process_file` after success.
