### 2026-04-04 erlingdevold/DINO-pipeline#14 @erlingdevold
Fix in c030b8e. All modules now use append mode when logging
