### 2026-04-04 erlingdevold/DINO-pipeline#17 @erlingdevold
Address by c030b8e. `MAX_WORKERS=os.cpu_count()` default.
