### 2026-04-04 erlingdevold/DINO-pipeline#18 @erlingdevold
Fix in c030b8e. `mask = None` at top of frequency loop iter prevents NaeError.
