### 2026-04-04 erlingdevold/DINO-pipeline#19 @erlingdevold
Fixed in c030b8e. Model loaded once in `consume_dir`, dead code removed, duplicate import removed.
