### 2026-04-04 erlingdevold/DINO-pipeline#15 @erlingdevold
Clarification: stage 3 is parallelized via batched GPU inference (not process pools). Multiple images are stacked into a single tensor and forwarded together, which is the standard approach for GPU workloads.
BATCH_SIZE controls this. See inference/inspect_attention.py.
CUDA OOM falls back to CPU automatically.
