### 2026-04-04 erlingdevold/DINO-pipeline#11 @erlingdevold
Added in 64d7787. Unit tests, CI runs on push. Asserts output correctness.
