### 2026-04-04 erlingdevold/DINO-pipeline#17 @erlingdevold
Also fixed in 64d7787, docker-compose.yml had max_workers= and DEVICE= (empty) which crashed python code.
