### 2026-04-04 erlingdevold/DINO-pipeline#7 @erlingdevold
8ca0665.  writing fixes, added benchmark table.
