### 2026-04-04 erlingdevold/DINO-pipeline#8 @erlingdevold
8ca0665. HDF%/netCDF dependency documented in README and docs/ with brew/apt installs.
