### 2026-04-04 erlingdevold/DINO-pipeline#9 @erlingdevold
created a mkdocs-material site with install,usage and gallery, should be deployed to gh-pages.
