### 2026-04-04 erlingdevold/DINO-pipeline#16 @erlingdevold
Addressed in 8ca0665. Replaced 'HPC machines' with 'laptop to multi-core servers'. README notes parallelism is intra-node only with no cluster/scheduler integration.
