### 2026-04-04 erlingdevold/DINO-pipeline#10 @erlingdevold
70c33d9. A full docker pipeline benchmark on 64 EK80 files is developed. Scales with MAX_WORKERS for CPU-bound stages. (8.5 s/file at 8 workers on i7-12700 + rtx3090 (~23 hr/TB) reproducible via bash download_bench_data.sh && uv run benchmark.py.
