### 2026-04-04 da4ac0f
feat: add benchmark pipeline, polish code and docs
Benchmark: download script, docker-compose override,
full Docker pipeline benchmark varying MAX_WORKERS. Answers issue #10.
Also: fix is_file_ready crash, remove dead code,
expand gallery with per-frequency echograms,
add Performance section to paper, strengthen test assertions.
Fix CI pip install (split torch from PyPI packages).
