+++ benchmark.py
@@ -1,140 +1,247 @@
-# %%
 """
-Benchmark script for EchoFlow pipeline stages.
+EchoFlow full-pipeline benchmark.
 
-Generates a markdown-formatted throughput table for the README.
-Run with: python benchmark.py
+Runs all three stages (raw -> preprocessing -> inference) inside Docker,
+varying MAX_WORKERS to measure throughput and scaling.
 
-Requires a sample .nc file in data/raw_consumer/ and corresponding
-PNG output in data/preprocessing/ for stages 2 and 3.
+Prerequisites:
+    - Docker and docker compose installed
+    - Benchmark data downloaded: bash download_bench_data.sh
+    - Docker images built: docker compose build
+
+Usage:
+    python benchmark.py [--data-dir data/benchmark/input] [--workers 1,2,4,8]
 """
 
 import os
 import platform
+import subprocess
 import time
+from dataclasses import dataclass
 from pathlib import Path
 
-import numpy as np
-import xarray as xr
+COMPOSE_FILES = ["docker-compose.yml", "docker-compose.benchmark.yml"]
 
 
 def get_machine_info():
-    """Print machine specs."""
-    import torch
-
+    """Gather machine specs for the report header."""
     lines = [
         f"- **OS:** {platform.system()} {platform.release()}",
         f"- **CPU:** {platform.processor() or platform.machine()}",
         f"- **Cores:** {os.cpu_count()}",
-        f"- **GPU:** {'CUDA ' + torch.cuda.get_device_name(0) if torch.cuda.is_available() else 'None (CPU only)'}",
     ]
+    try:
+        gpu = (
+            subprocess.check_output(
+                ["nvidia-smi", "--query-gpu=name", "--format=csv,noheader"],
+                text=True,
+                timeout=5,
+            )
+            .strip()
+            .split("\n")[0]
+        )
+        lines.append(f"- **GPU:** {gpu}")
+    except (FileNotFoundError, subprocess.SubprocessError):
+        lines.append("- **GPU:** None (CPU only)")
     return "\n".join(lines)
 
 
-def create_synthetic_nc(tmp_dir, n_files=1):
-    """Create synthetic .nc files for benchmarking."""
-    tmp_dir = Path(tmp_dir)
-    tmp_dir.mkdir(parents=True, exist_ok=True)
-    files = []
-    rng = np.random.default_rng(42)
+def count_files(directory, pattern):
+    """Count files matching glob pattern in directory."""
+    return len(list(Path(directory).rglob(pattern)))
 
-    for i in range(n_files):
-        freqs = [38000.0, 70000.0]
-        n_pings = 200
-        n_depths = 500
-        ping_time = np.array(
-            [np.datetime64("2024-01-01") + np.timedelta64(j, "s") for j in range(n_pings)]
+
+def run_stage(service, env_overrides):
+    """Run a docker compose service with -e flags and return elapsed seconds."""
+    cmd = ["docker", "compose"]
+    for f in COMPOSE_FILES:
+        cmd.extend(["-f", f])
+    cmd.extend(["run", "--rm"])
+    for k, v in env_overrides.items():
+        cmd.extend(["-e", f"{k}={v}"])
+    cmd.append(service)
+
+    start = time.perf_counter()
+    result = subprocess.run(cmd, capture_output=True, text=True)
+    elapsed = time.perf_counter() - start
+
+    if result.returncode != 0:
+        print(f"  WARNING: {service} exited with code {result.returncode}")
+        if result.stderr:
+            print(result.stderr[-500:])
+
+    return elapsed
+
+
+def clean_dir(path, protected):
[…]
