+++ inference/requirements.txt
@@ -1,5 +1,3 @@
-torch
 matplotlib
 numpy
 pillow
-torchvision

[…]
