### 2026-04-04 #9240 @erlingdevold
Hi @nkrush, all issues should be addressed now. Each issue should have a comment linking to the relevant commit. i have used the gh CLI api, to write comments, but will go over them online.

Most significant piece of work is the benchmark, detailing the performance.
Let me know if anything needs further work.
