### 2026-04-04 erlingdevold/DINO-pipeline#10 @subhk
@erlingdevold Thanks! I’ll get back to this within a week.
