### 2026-04-07 erlingdevold/DINO-pipeline#11 @nkrusch
I have checked this test suite and believe the JOSS test suite requirement is met satisfactorily.

I do have some minor comments, but they do not affect my satisfactory opinion.
1. I recover the test dependencies from the ci workflow file; these are better handled with a dev-requirements file.
2. The unit tests pass, but I see 3 warnings. The same warning appear in the [actions output](https://github.com/erlingdevold/EchoFlow/actions/runs/23975206193/job/69931060860), so I won't repeat here.

I defer to @subhk to close this issue.
