### 2026-04-11 erlingdevold/DINO-pipeline#11 @subhk
Thanks a lot, @nkrusch . It looks good now.

One suggestion: @erlingdevold, you may want to consider adding pytest-cov (or a similar tool) to the CI pipeline so that test coverage is visible to contributors.
