### 2026-04-12 #9240 @nkrusch
@erlingdevold Please first complete the two outstanding issues.

After that, I request that you complete the following steps.

1. Make a (new) tagged release of the software.
1. Archive the reviewed software in Zenodo, Figshare, or similar archival service.
1. Check the archival deposit has the correct metadata. This includes:
   - title (should match the paper)
   - author list (should match the paper)
   - author's ORCID (should match the paper)
   - license (should match the repository)
   - release version tag should match the archive version name
      For example, git tag v1.0 would have archive name v1.0 (displays as "Version v1.0" on Zenodo).
1. List here, as a comment:
   - the version tag
   - the archive DOI

I can then move forward with recommending accept for the submission.
