### 2026-04-12 #9240 @erlingdevold
Both outstanding issues now committed.

1. Make a (new) tagged release of the software.
 -- tagged v1.0.1 and pushed.
2. Archive the reviewed software in Zenodo, Figshare, or similar archival service.
-- v1.0.1 archived at Zenodo with doi. [10.5281/zenodo.19534258](https://doi.org/10.5281/zenodo.19534258)
Check the archival deposit has the correct metadata. This includes:
title: EchoFlow
- Correct
author list: Erling Devold
- Correct
author's ORCID : 0009-0000-0949-5992
- Correct
license: MIT
- Correct
release version tag should match the archive version name
- GitHub release version: v1.0.1
- Version v1.0.1 w/ doi https://doi.org/10.5281/zenodo.19534258

List here, as a comment:
the version tag
the archive DOI

EDIT:
Issues not resolved, solution is committed and marked in issues.
