### 2026-04-12 #9240 @nkrusch
@erlingdevold please update the Zenodo title field to match the paper title:
> EchoFlow: End-to-end self-supervised sonar-image pipeline
