### 2026-04-21 #9240 @crvernon
🥳 Congratulations on your new publication @erlingdevold ! Many thanks to @nkrusch for editing and reviewing and @subhk for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
