### 2026-04-21 #9240 @erlingdevold
Thanks all for a great review! @nkrusch, @subhk Thanks for your patience, this has been a great experience.
I'll recommend the journal to my peers at SINTEF! What a blast it's been doing a scientific peer review in the format of a pull request!
